refactor(shared): create Shared/ BC + migrate AbstractBatchHandler/Metrics/Maintenance
Bootstrap Shared/ bounded context with the four DDD layers and move three
cross-cutting modules out of top-level src/:

- src/Messaging/AbstractBatchHandler.php
  -> src/Shared/Application/AbstractBatchHandler.php
- src/Observability/Http/MetricsController.php
  -> src/Shared/Infrastructure/Metrics/MetricsController.php
- src/Maintenance/{DatabaseResetCommand,TenantAuditCommand}.php
  -> src/Shared/Infrastructure/Maintenance/

Empty src/Messaging, src/Observability, src/Maintenance directories
removed. Namespaces updated in moved files; the only consumer of the
batch handler base (Catalog\Application\Handler\RebuildAttributesIndexedHandler)
and the benchmark phpdoc reference now point at the Shared namespace.

Refs #151
