+++ apps/api/src/Shared/Infrastructure/Metrics/MetricsController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Observability\Http;
+namespace App\Shared\Infrastructure\Metrics;
 
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
